+++ src/ui/menu/room/download_level_popup.cpp
@@ -1,6 +1,7 @@
 #include "download_level_popup.hpp"
 
 #include <util/ui.hpp>
+#include <util/gd.hpp>
 
 using namespace geode::prelude;
 
@@ -54,31 +55,9 @@ void DownloadLevelPopup::loadLevelsFailed(char const* p0) {
 void DownloadLevelPopup::setupPageInfo(gd::string p0, char const* p1) {}
 
 void DownloadLevelPopup::levelDownloadFinished(GJGameLevel* level) {
-    // thank you cvolton :D
-    // this is needed so the level appears at the top of the saved list (unless Manual Level Order is enabled)
-
-    auto* levels = GameLevelManager::get()->m_onlineLevels;
-
-    bool putAtLowest = GameManager::get()->getGameVariable("0084");
-
-    int idx = 0;
-    for (const auto& [k, level] : CCDictionaryExt<gd::string, GJGameLevel*>(levels)) {
-        if (putAtLowest) {
-            idx = std::min(idx, level->m_levelIndex);
-        } else {
-            idx = std::max(idx, level->m_levelIndex);
-        }
-    }
-
-    if (putAtLowest) {
-        idx -= 1;
-    } else {
-        idx += 1;
-    }
-
-    level->m_levelIndex = idx;
-
     this->onClose(this);
+
+    util::gd::reorderDownloadedLevel(level);
     util::ui::switchToScene(LevelInfoLayer::create(level, false));
 }