+++ src/util/gd.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <Geode/Bindings.hpp>
+
+namespace util::gd {
+    void reorderDownloadedLevel(GJGameLevel* level);
+}